\name{neftali.barrero-package}
\alias{neftali.barrero-package}
\alias{neftali.barrero}
\docType{package}
\title{

Devuelve el Cuadrado o el Cubo de un número
~~ CUADRADO_CUBO ~~
}
\description{
-tiene como parametro de entrada un número ydevuelve el cuadrado del mismo.

}
\details{
\tabular{ll}{
Package: \tab neftali.barrero\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-15\cr
License: \tab NO\cr
}
cuadrado: devuevle el cuadrado del parametro de netrada X
cubo: devuevle el cubo del parametro de netrada X
}
\author{
NEFTALI BARRERO PARRA
Ing. Informático
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
cuadrado(3)  
9

cubo(2)

8
}
